write print controller

+++ internal/task/controller/task/delete.go
@@ -9,6 +9,9 @@ func (h *Handler) Delete(args []string) (string, error) {
 	if len(args) == 0 {
 		return "", ErrFewArguments
 	}
+	if len(args) > 1 {
+		return "", ErrALotOFArguments
+	}
 	id, _ := strconv.Atoi(args[0])
 
 	err := h.service.Delete(id)
